+++ WordPress/src/main/AndroidManifest.xml
@@ -161,7 +161,7 @@
             android:label="@string/add_new_category"
             android:theme="@style/Theme.AppCompat.Light.Dialog" />
         <activity
-            android:name=".ui.photopicker.MediaPreviewActivity"
+            android:name=".ui.media.MediaPreviewActivity"
             android:theme="@style/Theme.AppCompat.NoActionBar" />
         <activity
             android:name=".editor.legacy.EditLinkActivity"
